requirements.txt
fastapi==0.95.2
python-dotenv==1.0.0
uvicorn==0.22.0

# Build
openapi-spec-validator==0.5.6

# Testing
pytest==7.3.1
pytest-cov==4.1.0
starlette[full]==0.27.0
